Rename ProseMirror component
