+++ src/components/editor/index.js
@@ -1,2 +1,2 @@
-export * from './container'
-export * from './toolbar'
+export * from './container.js'
+export * from './toolbar.js'
